Revert "Include debian-installer components into the mirroring."
This reverts commit 591a52e67a40e8fee2e446de7c151845a03268a9.

The standard way to include the installation components is via the
mirror spec line, e.g.x

deb-amd64 http://us.archive.ubuntu.com/ubuntu wily main restricted universe multiverse main/debian-installer

This is a universal syntax, used by everybody in the same way. I think
it would be better not for apt-mirror to fetch things I didn't ask for,
and to simply continue it's current behavior of doing what it is asked
to do.

fixes #64