Anything downloaded should be kept
Do the skipclean logic in only one place. The logic is pretty simple -
if we downloaded a file, then we *must* keep it, so it is logical to add
it to `add_url_to_download`.

Also dropping the logic to keep the uncompressed index files as per the
previous commit we do not need to host them.