Refactor: Anything downloaded should be kept (#19)
Refactor this skipclean logic to only do it in one place. The logic is
pretty simple - if we decided to download a file, then we *must* keep
it.

There is no need for the copy-pasta.

Not keeping the logic to keep the uncompressed version of compressed
files, because if we did not download the uncompressed files, there is
no reason why we should be trying to keep them.