Only download compressed index files
I cannot find a single repository serving the uncompressed index files.

Furthermore, per the documentation at
https://wiki.debian.org/DebianRepository/Format#Compression_of_indices

    Servers should offer only xz compressed files, except for the
    special cases listed above.

Therefore we should only download the compressed files, and only serve
the compressed files, and all clients are expected to understand them.

The uncompressed versions are documented in the Release file, but we can
safely ignore them and only get the compressed ones instead.