wget timeouts excessive
In recent testing (as of this commit) 3 items were unavailable from deb.debian.org/debian (site unreachable: ERR_CONNECTION_RESET):
```
http://deb.debian.org/debian/pool/main/libc/libcorona-perl/libcorona-perl_0.1004-4_all.deb
http://deb.debian.org/debian/pool/main/n/nicovideo-dl/nicovideo-dl_0.0.20190126-1.1_all.deb
http://deb.debian.org/debian/pool/main/r/r-cran-covid19us/r-cran-covid19us_0.1.7-1_all.deb
```
The result is a connection timeout after the default wget timeout of 900 second - per link. Coupled with the '-t 5' (--retries) we had a stall in operation for 1h15m. This was during concurrent threads. if the nthreads was 1 this would have resulted in 3h45m delay in processing.

The solution was to reduce connection timeout to 60 seconds, and reduce the retries to only once.
-T,  (--timeout=SECONDS)  Same as setting all these to the same timeout: --dns-timeout, --connect-timeout, --read-timeout
-t, (--tries=NUMBER)  Set number of retries to NUMBER. 0=unlimited

These values are acceptable with today's internet, and if a file fails during a run, it will get picked up next time around, assuming the problem has been resolved.

This could also be a candidate for a configurable option(s) in the config file.