Resume mirroring after an interrupted state
If wget gets interrupted the currently downloading file will be
incomplete, and will have a very recent timestamp.

In this situation, wget will not re-download the file since it will
appear to be more recent than the file it is trying to download.

"--no-if-modified-since" works around this problem by ignoring the
timestamp of the local file.